Move server to seperate file
